Improved param extraction and fixed memory leaks.
This commit includes an additional helper function which parses
the variable and clause parameters from the input file. It also
defines some constants that are used to print out the output
messages (for consistency). Finally, I added the '-g' flag to
the Makefile so that GDB can be used to quickly identify where
errors are occuring.